+++ jobs/provisional-approval/src/provisional_approval/job.py
@@ -23,6 +23,7 @@ from strr_api.models.rental import Registration
 from strr_api.services import ApprovalService
 from structured_logging import StructuredLogging
 from enum import Enum
+from sqlalchemy import or_
 
 from provisional_approval.config import CONFIGURATION
 
@@ -54,8 +55,11 @@ def get_applications_in_full_review_status(app):
         Application.query.filter(
             Application.status == Application.Status.FULL_REVIEW,
             Application.registration_type == Registration.RegistrationType.HOST,
-            # StrrRequirement.PR_EXEMPT_FRACTIONAL_OWNERSHIP.value
-            # Application.registration.has("prExemptionType" != "FRACTIONAL_OWNERSHIP")
+            or_(
+                Application.application_json["registration"]["prExemptionType"].astext.is_(
+                    None),
+                Application.application_json["registration"]["prExemptionType"].astext != "FRACTIONAL_OWNERSHIP"
+            )
         )
         .order_by(Application.id)
         .limit(int(app.config.get("BATCH_SIZE")))
